refactor: create domains
