+++ src/ctxs/fasta/application/mod.rs
@@ -1,3 +1,6 @@
-pub mod edit;
-pub mod make;
+pub mod chain_operations;
+pub mod cut;
+pub mod format;
+pub mod generate;
+pub mod to_aminoacids;
 pub mod view;
