+++ src/ctxs/fasta/infrastructure/mod.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use crate::shared::infrastructure::CommonWriteFormat;
+use crate::ctxs::shared::infrastructure::CommonWriteFormat;
 
 use super::domain::fasta::Fasta;
