+++ src/ctxs/mod.rs
@@ -1,4 +1,6 @@
 pub mod aminoacid;
 pub mod codon;
+pub mod dna;
 pub mod fasta;
 pub mod protein;
+pub mod shared;
